requirements.txt
yt-dlp
whisper-openai
torch
